chore: remove compression, refactor code
