+++ apps/backend/package.json
@@ -50,7 +50,6 @@
     "amqplib": "^0.10.5",
     "auth-header": "^1.0.0",
     "axios": "^1.7.9",
-    "compression": "^1.7.5",
     "convict": "^6.2.4",
     "cookies": "^0.9.1",
     "cors": "^2.8.5",
@@ -91,7 +90,6 @@
   "devDependencies": {
     "@types/amqplib": "^0.10.6",
     "@types/auth-header": "^1.0.6",
-    "@types/compression": "^1.7.5",
     "@types/convict": "^6.1.6",
     "@types/cookies": "^0.9.0",
     "@types/cors": "^2.8.17",
